requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0,<2.0
opencv-python>=4.8.0
PyYAML>=6.0
tqdm>=4.66.0
Flask>=3.0.0
albumentations>=1.4.0,<2.0
